final-pjt/final_pjt_back/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
colorama==0.4.6
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.1
fonttools==4.55.0
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.7
jiter==0.7.1
multidict==6.1.0
openai==0.28.0
propcache==0.2.0
pydantic==2.10.1
pydantic_core==2.27.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
zipp==3.21.0
yarl==1.18.0

